Make login and signup pages navigate home
